initial cleanup.
